Address PR feedback
Here, `ServiceUpdatePeriod` is made a bit more typesafe and expressive
by having it wrap a `time::Duration` directly.

Due to the mismatches between `std::time::Duration` and
`time::Duration`, the underlying parsing is now expressed in terms of
`u64`, rather than a `u64`, which could exceed the range of
`time::Duration`. A `u32` gives a maximum range of 49+ days, which is
quite generous for this usecase.

The `Periodic` trait has been modified accordingly.

Signed-off-by: Christopher Maier <cmaier@chef.io>